spring-data-jdbc 적용. embedded Id 미적용

+++ recruit/src/main/resources/config/liquibase/sql/sample_init.sql
@@ -1,4 +1,4 @@
 create table sample (
-    id int primary key,
+    id bigint primary key,
     name varchar(255)
 );
